// components/student/studentAttestationList/studentAttestationListStyle.scss.d.ts
export type Styles = {
  'attestationBtnItem': string;
  'attestationBtnNotAvailable': string;
  'attestationItemContainer': string;
  'attestationItemContentContainer': string;
  'attestationName': string;
  'linkShowAttestationTests': string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;